+++ android/envoy/src/main/kotlin/org/greatfire/envoy/EnvoyState.kt
@@ -1,5 +1,7 @@
 package org.greatfire.envoy
 
+import org.greatfire.envoy.transport.Transport
+
 import android.content.Context
 import android.util.Log
 import IEnvoyProxy.IEnvoyProxy // Go library
