+++ android/envoy/src/main/kotlin/org/greatfire/envoy/transport/CronetEnvoyTransport.kt
@@ -1,4 +1,7 @@
-package org.greatfire.envoy
+package org.greatfire.envoy.transport
+
+import org.greatfire.envoy.CronetNetworking
+import org.greatfire.envoy.EnvoyServiceType
 
 import android.content.Context
 import android.net.Uri
