+++ android/envoy/src/main/kotlin/org/greatfire/envoy/transport/CronetMasqueTransport.kt
@@ -1,5 +1,6 @@
-package org.greatfire.envoy
+package org.greatfire.envoy.transport
 
+import org.greatfire.envoy.EnvoyServiceType
 
 import IEnvoyProxy.IEnvoyProxy
 import android.content.Context
