+++ android/envoy/src/main/kotlin/org/greatfire/envoy/transport/CronetProxyTransport.kt
@@ -1,4 +1,6 @@
-package org.greatfire.envoy
+package org.greatfire.envoy.transport
+
+import org.greatfire.envoy.EnvoyServiceType
 
 import android.content.Context
