+++ android/envoy/src/main/kotlin/org/greatfire/envoy/transport/OkHttpEnvoyTransport.kt
@@ -1,11 +1,13 @@
-package org.greatfire.envoy
+package org.greatfire.envoy.transport
+
+import org.greatfire.envoy.EnvoyServiceType
 
 import android.content.Context
 import android.net.Uri
 import android.util.Log
 import okhttp3.Request
 
-class OkHttpProxyTransport(url: String) : Transport(EnvoyServiceType.OKHTTP_PROXY, url) {
+class OkHttpEnvoyTransport(url: String) : Transport(EnvoyServiceType.OKHTTP_ENVOY, url) {
 
     override suspend fun startTest(context: Context): Boolean {
         val host = Uri.parse(testUrl).host
@@ -15,17 +17,26 @@ class OkHttpProxyTransport(url: String) : Transport(EnvoyServiceType.OKHTTP_PROX
         }
 
         // XXX cache param, this is hacky :)
-        // this nshould be updated to use the same checksum param
+        // this should be updated to use the same checksum param
         // that the C++ patches used to use
         val t = System.currentTimeMillis()
-        val url = proxyUrl.toString() + "?test=" + t
+        val tempUrl = url + "?test=" + t
         val request = Request.Builder()
-            .url(url)
+            .url(tempUrl)
             // .head()  // a HEAD request is enough to test it works
             .addHeader("Url-Orig", testUrl)
             .addHeader("Host-Orig", host)
             .build()
 
-        return runTest(request, null)
+        val temp = runTest(request, null)
+        if (temp == true && this.proxyUrl.isNullOrEmpty()) {
+            // the Envoy proxy URL needs to be copied to proxyUrl
+            this.proxyUrl = url
+        }
+        return temp
+    }
+
+    override fun stopService() {
+        // no service to stop
     }
 }
