+++ android/envoy/src/main/kotlin/org/greatfire/envoy/transport/V2SrtpTransport.kt
@@ -1,4 +1,7 @@
-package org.greatfire.envoy
+package org.greatfire.envoy.transport
+
+import org.greatfire.envoy.EnvoyConnectionTests
+import org.greatfire.envoy.EnvoyServiceType
 
 import android.net.Uri
 import android.net.UrlQuerySanitizer
@@ -7,10 +10,10 @@ import android.util.Log
 import IEnvoyProxy.IEnvoyProxy
 import android.content.Context
 
-class V2WechatTransport(url: String) : Transport(EnvoyServiceType.V2WECHAT, url) {
+class V2SrtpTransport(url: String) : Transport(EnvoyServiceType.V2SRTP, url) {
 
     override suspend fun startTest(context: Context): Boolean {
-        val addr = startService()
+        var addr = startService()
 
         if (addr == "") {
             // The go code doesn't handle failures well, but an empty
@@ -20,7 +23,7 @@ class V2WechatTransport(url: String) : Transport(EnvoyServiceType.V2WECHAT, url)
         }
 
         proxyUrl = "socks5://$addr"
-        Log.d(TAG, "testing V2Ray WeChat at ${proxyUrl}")
+        Log.d(TAG, "Testing V2Ray SRTP ${proxyUrl}")
 
         val res = testStandardProxy(Uri.parse(proxyUrl), testResponseCode)
         if (res == false) {
@@ -31,11 +34,11 @@ class V2WechatTransport(url: String) : Transport(EnvoyServiceType.V2WECHAT, url)
 
     override suspend fun startService(): String {
         if (serviceRunning) {
-            Log.e(TAG, "Tried to start V2ray Wechat when it was already running")
+            Log.e(TAG, "Tried to start V2ray Srtp when it was already running")
             return ""
         }
 
-        Log.d(TAG, "Starting service for V2ray Wechat")
+        Log.d(TAG, "Starting service for V2ray Srtp")
 
         serviceRunning = true
 
@@ -46,12 +49,8 @@ class V2WechatTransport(url: String) : Transport(EnvoyServiceType.V2WECHAT, url)
             it.v2RayServerPort = server.port.toString()
             it.v2RayId = getV2RayUuid(url)
 
-            val host = server.host
-            val port = server.port.toString()
-            val uuid = getV2RayUuid(url)
-
-            it.start(IEnvoyProxy.V2RayWechat, "")
-            val addr = it.localAddress(IEnvoyProxy.V2RayWechat)
+            it.start(IEnvoyProxy.V2RaySrtp, "")
+            val addr = it.localAddress(IEnvoyProxy.V2RaySrtp)
             EnvoyConnectionTests.isItUpYet(addr)
             return addr
         }
@@ -64,7 +63,7 @@ class V2WechatTransport(url: String) : Transport(EnvoyServiceType.V2WECHAT, url)
         // this is called to stop unused services
 
         state.iep?.let {
-            it.stop(IEnvoyProxy.V2RayWechat)
+            it.stop(IEnvoyProxy.V2RaySrtp)
         }
     }
